backend/requirements.txt
# Web
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart

# Data
polars>=1.0.0
fastexcel
openpyxl
xlsxwriter

# DB
sqlmodel>=0.0.21
alembic>=1.13.0
asyncpg
databases
psycopg2-binary

# Cache / Queue
redis>=5.0.0
celery>=5.4.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# AI
openai>=1.30.0
apscheduler>=3.10.4
edge-tts>=6.1.9
faster-whisper>=1.0.0
httpx>=0.27.0

# Config
python-dotenv>=1.0.0

# Dev
pytest
pytest-asyncio
httpx
pydantic-settings
